+++ src/components/ui/Beta.tsx
@@ -11,7 +11,7 @@ export default function Beta({ className }: BetaProps): ReactElement {
     <span
       className={cn(
         'text-[10px] text-black uppercase px-1 py-0.5 select-none',
-        'max-w-[29px] ml-1 bg-dynamic brightness-125 rounded-md align-top',
+        'max-w-[29px] ml-1 bg-dynamic brightness-125 rounded-md',
         className,
       )}
     >