+++ src/types/navigation.ts
@@ -1,4 +1,4 @@
-import type { Dispatch, SetStateAction } from 'react'
+import type { ComponentType, Dispatch, SetStateAction } from 'react'
 
 export type CustomListType = 'cardFarmingList' | 'achievementUnlockerList' | 'autoIdleList' | 'favoritesList'
 export type ActivePageType =
@@ -9,6 +9,7 @@ export type ActivePageType =
   | 'settings'
   | `customlists/${string}`
   | 'tradingCards'
+  | 'chat'
 export type CurrentTabType = 'achievements' | 'statistics'
 export type CurrentSettingsTabType = 'general' | 'card-farming' | 'achievement-unlocker' | 'debug'
 
@@ -25,9 +26,10 @@ export interface SidebarItem {
   id: string
   page: ActivePageType
   title: string
-  icon: React.ComponentType<{ fontSize?: number; className?: string }>
+  icon: ComponentType<{ fontSize?: number; className?: string }>
   shouldShow?: boolean
   isActive?: boolean
   customClassName?: string
   hasDivider?: boolean
+  isBeta?: boolean
 }